requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.31.0
transformers>=4.40.0
accelerate>=0.30.0
lpips>=0.1.4
kornia>=0.7.0
einops>=0.7.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
